Transformer-based/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "anomaly-attack-classifier"
version = "0.2.0"
description = "Transformer-based attack-type classification for SWAT/WADI-like datasets"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "scikit-learn>=1.3",
  "pyyaml>=6.0",
  "openpyxl>=3.1",
  "torch>=2.2",
]

[tool.setuptools]
packages = ["anomaly_attack"]
